requirements.txt
sqlalchemy
fastapi
uvicorn
streamlit
pydantic
requests
python-multipart
pyyaml
python-dotenv
pandas